重构、新功能 在service层判断了节点不存在等问题（通过返回false或null等方式）最终在controller层用ResponseVO封装处理；我认为现在的entity与relationship应该被称为javaBean而不是pojo于是改名了；在entity中添加了relatesTo来记录与之相关的实体与关系，并添加了一些简单的方法来获得它们；修改了添加关系、删除关系、删除节点的内部逻辑（它们都需要更新relatesTo）

+++ src/main/java/com/example/coin/service/impl/EntityServiceImpl.java
@@ -3,10 +3,11 @@ package com.example.coin.service.impl;
 import com.example.coin.DAO.EntityRepository;
 import com.example.coin.javaBeans.Entity;
 import com.example.coin.service.EntityService;
+import com.example.coin.service.RelationshipService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,7 +16,7 @@ public class EntityServiceImpl implements EntityService {
     @Autowired
     private EntityRepository entityRepository;
     @Autowired
-    private MongoTemplate mongoTemplate;
+    private RelationshipService relationshipService;
 
     /**
      * 将关系节点持久化到数据库中
@@ -32,8 +33,16 @@ public class EntityServiceImpl implements EntityService {
      * @param id
      */
     @Override
-    public void deleteEntityById(String id) {
+    public boolean deleteEntityById(String id) {
+        Entity entity = findEntityById(id);
+        if(entity == null) return false;
+        //删除与节点相关的所有关系
+        HashSet<String> associatedRelationships = entity.associatedRelationships();
+        for(String relsId : associatedRelationships){
+            relationshipService.deleteRelationById(relsId);
+        }
         entityRepository.deleteById(id);
+        return true;
     }
 
     /**
@@ -78,6 +87,8 @@ public class EntityServiceImpl implements EntityService {
     @Override
     public void deleteAllEntities() {
         entityRepository.deleteAll();
+        //删除所有实体节点的同时删除所有关系节点
+        relationshipService.deleteAllRelationships();
     }
 
 }
